+++ src/lib/Types.sol
@@ -3,8 +3,8 @@ pragma solidity 0.8.20;
 
 library Types {
     struct BlockHeaderProof {
-        uint128 treeId;
-        uint256 mmrTreeSize;
+        uint256 treeId;
+        uint128 mmrTreeSize;
         uint256 blockNumber;
         uint256 blockProofLeafIndex;
         bytes32[] mmrPeaks;
